requirements.txt
transformers
torch
langchain
langchain-community
langchain-core
langchain-google-genai
python-dotenv
requests
streamlit